Restriction on Promo CRUD route + comment ordering on deals

+++ src/Controller/PromoController.php
@@ -2,12 +2,16 @@
 
 namespace App\Controller;
 
+use App\Entity\Comment;
 use App\Entity\Promo;
+use App\Form\CommentType;
 use App\Form\PromoType;
 use App\Repository\PromoRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
@@ -30,6 +34,7 @@ class PromoController extends AbstractController
 
     /**
      * @Route({"en": "/new", "fr": "/ajouter"}, name="promo_new", methods={"GET","POST"})
+     * @IsGranted("IS_AUTHENTICATED_REMEMBERED")
      */
     public function new(Request $request): Response
     {
@@ -53,20 +58,39 @@ class PromoController extends AbstractController
     }
 
     /**
-     * @Route("/{slug}", name="promo_show", methods={"GET"})
+     * @Route("/{slug}", name="promo_show", methods={"GET","POST"})
      */
-    public function show(Promo $promo): Response
+    public function show(Request $request, Promo $promo): Response
     {
+        $newComment = new Comment();
+        $newComment->setDeal($promo);
+        $commentForm = $this->createForm(CommentType::class, $newComment);
+        $commentForm->handleRequest($request);
+
+        if ($commentForm->isSubmitted() && $commentForm->isValid()) {
+
+            $entityManager = $this->getDoctrine()->getManager();
+            $newComment->setAuthor($this->getUser());
+            $entityManager->persist($newComment);
+            $entityManager->flush();
+        }
+
         return $this->render('pages/promo/show.html.twig', [
             'promo' => $promo,
+            'commentForm' => $commentForm->createView(),
         ]);
     }
 
     /**
      * @Route({"en": "/{id}/edit", "fr": "/{id}/edition"}, name="promo_edit", methods={"GET","POST"})
+     * @IsGranted("IS_AUTHENTICATED_REMEMBERED")
      */
     public function edit(Request $request, Promo $promo): Response
     {
+        if ($this->getUser()->getId() != $promo->getAuthor()->getId() || in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            throw new AccessDeniedHttpException();
+        }
+
         $form = $this->createForm(PromoType::class, $promo);
         $form->handleRequest($request);
 
@@ -83,10 +107,15 @@ class PromoController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="promo_delete", methods={"POST"})
+     * @Route("/{id}/delete", name="promo_delete", methods={"POST"})
+     * @IsGranted("IS_AUTHENTICATED_REMEMBERED")
      */
     public function delete(Request $request, Promo $promo): Response
     {
+        if ($this->getUser()->getId() != $promo->getAuthor()->getId() || in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            throw new AccessDeniedHttpException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$promo->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($promo);
